requirements.txt
flask==3.0.0
huggingface-hub==0.20.2
ultralytics==8.1.0
pillow==10.2.0
opencv-python-headless==4.9.0.80
numpy==1.26.3
torch==2.1.2
torchvision==0.16.2
gunicorn==21.2.0
